\name{signalextraction-package}
\alias{signalextraction-package}
\alias{signalextraction}
\docType{package}
\title{Direct Filter Approach}
\description{The Direct Filter Approach (DFA) provides efficient
  estimates of signals at the current boundary of time series in
  real time. For that purpose, one-sided ARMA-filters are computed by
  minimizing customized error criteria. The DFA can be used for
  estimating either the level or turning-points of a series, owing to
  the fact that that both criteria are incongruent and cannot be
  optimized simultaneously. In the context of real time turning-point
  detection, various risk profiles can be operationalized, which account
  for the speed and/or the reliability of the one-sided filter.}
\details{
\tabular{ll}{Package: \tab signalextraction    \cr
             Type:    \tab Package             \cr
             Version: \tab 2.0.0               \cr
             Date:    \tab 2007-03-08          \cr
             License: \tab GPL 2.0 or newer    \cr}}
\author{Marc Wildi & Marcel Dettling, <[wia/dem]@zhwin.ch>}
\references{Marc Wildi, \emph{Real-Time Signal Extraction (Beyond
    Maximum Likelihood Principles)}, Springer. To appear in 2007.}
\keyword{package}

