\name{x}
\alias{x}
\docType{data}
\title{Business Survey Data}
\description{In order to anticipate movements of the GDP growth rate the
  Swiss Institute of Business Cycle Research (KOF-ETH) conducts a survey
  based on a panel comprising around 7000 companies (manufacturing,
  construction, retail, hotel, catering, wholesale, banks). The
  questionnaire can be viewed at \url{http://survey.kof.ethz.ch}. The
  Business Survey Data that are provided here are a selection of the
  sample which is used for construction of the leading indicator
  (KOF-Konjunkturbarometer).} 
\usage{data(x)} 
\format{It is a named numerical vector, where the names describe the
  time of measurement.}
\source{Swiss Institute for Business Cycle Research}
\references{Swiss Institute for Business Cycle Research, WEH D4, CH-8092
  Z\"urich. Contact person: Richard Etter, <ind@kof.ethz.ch>. See also
  \url{http://www.kof.ethz.ch/}.}
\examples{
data(x)
str(x)}
\keyword{datasets}
